+++ artifacts/contracts/Exchange/Game_Exchange.sol/Game_Exchange.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "../../../build-info/3a5bdaa1c2430d9593c50a941265d9c9.json"
+}
