--- artifacts/contracts/Items_Exchange/IGame_Items_Exchange.sol/IGame_Items_Exchange.dbg.json
@@ -1,4 +1,0 @@
-{
-  "_format": "hh-sol-dbg-1",
-  "buildInfo": "../../../build-info/754579e6ea18f4bb0f9302bec058bdca.json"
-}
